Update Driver.java
This is a Map/Reduce Maven build with dependencies for the Apache Hadoop Library.  This along with the other Map/Reduce files takes in two files of type .csv, and joins the files based on a certain key, the student ID.  It then sorts and groups the joined data sets based on student ID.  This will run on the hadoop ecosystem for Big Data purposes.